requirements.txt
PyYAML
whoosh
sentence-transformers
torch
faiss-cpu
numpy
mistralai
openai